+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.2.21] - 2026-04-07
+
+### Added 
+- Added Basic Festim support 
+- Refactored base class to include provider specific dictionaries
+
 ## [0.2.20] - 2026-04-06
 
 ### Added